backend/requirements.txt
Django==6.0.1
djangorestframework==3.14.0
django-filter==24.1
python-dotenv==1.0.0
qrcode[pil]==7.4.2
django-cors-headers==4.3.1
celery==5.3.4
celery-beat==2.5.0
redis==5.0.1